전체 크롤링 관련 모음 bucket 언젠가 쓸까 싶어서 남겨놓기 위한 브랜치...ㅎㅎ

+++ src/main/java/com/_02server/Device.java
@@ -5,7 +5,7 @@ public class Device {
     public String[] contents = new String[50];
     public String category;
     public double maximum_output; //최대 출력
-    public int wattPerhour; // Wh 있다면 여기 채우기
+    public double wattPerhour; // Wh 있다면 여기 채우기
     public int mAh; // mAh 있다면 여기 채우기
     public double volt = 3.7; // 기본 전압량 설정함.
     public String imageURL;
@@ -58,11 +58,11 @@ public class Device {
         this.maximum_output = maximum_output;
     }
 
-    public int getWattPerhour() {
+    public double getWattPerhour() {
         return wattPerhour;
     }
 
-    public void setWattPerhour(int wattPerhour) {
+    public void setWattPerhour(double wattPerhour) {
         this.wattPerhour = wattPerhour;
     }
